[YARN-474] - temporary fix until they fix issue: https://issues.apache.org/jira/browse/YARN-474?focusedCommentId=13615051&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-13615051

+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-resourcemanager/src/test/java/org/apache/hadoop/yarn/server/resourcemanager/scheduler/capacity/TestLeafQueue.java
@@ -1607,7 +1607,7 @@ public class TestLeafQueue {
     assertEquals(2, e.activeApplications.size());
     assertEquals(1, e.pendingApplications.size());
 
-    csConf.setDouble(CapacitySchedulerConfiguration
+    csConf.setFloat(CapacitySchedulerConfiguration
         .MAXIMUM_APPLICATION_MASTERS_RESOURCE_PERCENT,
         CapacitySchedulerConfiguration
         .DEFAULT_MAXIMUM_APPLICATIONMASTERS_RESOURCE_PERCENT * 2);
